Continuously send messages to create load
