+++ ethereum/package.json
@@ -19,6 +19,7 @@
     "build": "forge build -o build",
     "unit-test": "forge test -vvvvv",
     "integration-test": "bash shell-scripts/run_integration_tests.sh",
+    "load-test": "ENV=testnet ts-node ts-scripts/mockIntegrations/loadGeneration.ts",
     "typechain": "bash ../sdk/scripts/make_ethers_types.sh",
     "flatten": "mkdir -p node_modules/@poanet/solidity-flattener/contracts && cp -r contracts/* node_modules/@poanet/solidity-flattener/contracts/ && poa-solidity-flattener",
     "deployAndConfigureTilt": "ENV=tilt bash ./ts-scripts/shell/deployConfigureTest.sh",
