+++ ethereum/ts-scripts/config/testnet/contracts.json
@@ -48,23 +48,23 @@
   "mockIntegrations": [
     {
       "chainId": 6,
-      "address": "0x0979c1b2c04E0268a1DBF1800D5561354C719Ff9"
+      "address": "0xaC19a80C3409F8296043FC24F955790164Ccb15E"
     },
     {
       "chainId": 14,
-      "address": "0x1C6508C24C05f631C1771dcF064AB46e8c7ADFd0"
+      "address": "0xC02d04f9067Ede6fF4CaDC31233c795ec982eFd7"
     },
     {
       "chainId": 4,
-      "address": "0xFB29B4756F5F81991F621305f52BecF5fe0b2576"
+      "address": "0x3C1cd47E8D9aa563B238841B8D95c80d11E6B3A6"
     },
     {
       "chainId": 5,
-      "address": "0x62631C37a91f35e8F50A90e9A0cf09B3f0eae150"
+      "address": "0xf4092f14f74E119fae5343d08C3A253FB9D2e8d6"
     },
     {
       "chainId": 16,
-      "address": "0xc636a63001Ca975E409aa01E65e1982baaef41C2"
+      "address": "0x0896fB74CaBeDb603E3A6EdA36f40af00010Bf4c"
     }
   ]
 }
